refactor: move update credentials endpoint to project module
