+++ src/connection-request/connection-request.controller.ts
@@ -21,7 +21,6 @@ import {
   ConnectionDecisionDto,
   ConnectionRequestResponseDto,
   DatabaseTestDto,
-  UpdateCredentialsDto,
 } from './dto';
 import {
   ApiBadRequestResponse,
@@ -230,69 +229,6 @@ export class ConnectionRequestController {
       // handle rest - no code or an unknown code
       throw new InternalServerErrorException('Unexpected database error');
     }
-  }
-
-  @UseGuards(ResourceGuard)
-  @Scopes('view', 'edit', 'connect')
-  @Patch(':projectId/credentials')
-  @ApiOperation({
-    summary: 'Update connection credentials and retry connection',
-  })
-  @HttpCode(HttpStatus.NO_CONTENT)
-  @ApiNoContentResponse({
-    description: 'Credentials updated and connection retried',
-  })
-  @ApiBadRequestResponse({
-    description: 'Invalid project data for database operation',
-    content: {
-      'application/json': {
-        schema: { $ref: getSchemaPath(GenericErrorResponseDto) },
-        example: {
-          statusCode: 400,
-          message: 'Invalid project data for database operation',
-          error: 'DatabaseError',
-        },
-      },
-    },
-  })
-  @ApiNotFoundResponse({
-    description: 'Project not found',
-    content: {
-      'application/json': {
-        schema: { $ref: getSchemaPath(GenericErrorResponseDto) },
-        example: {
-          statusCode: 404,
-          message: 'Requested resource could not be found',
-          error: 'DatabaseError',
-        },
-      },
-    },
-  })
-  @ApiInternalServerErrorResponse({
-    description: 'Unexpected database error',
-    content: {
-      'application/json': {
-        schema: { $ref: getSchemaPath(GenericErrorResponseDto) },
-        example: {
-          statusCode: 500,
-          message: 'Database is temporarily unavailable',
-          error: 'DatabaseError',
-        },
-      },
-    },
-  })
-  async updateCredentials(
-    @Req() request: Request,
-    @CurrentUser() user: CurrentUserInfo,
-    @Param('projectId', ParseUUIDPipe) projectId: string,
-    @Body() dto: UpdateCredentialsDto,
-  ) {
-    return await this.connectionRequestService.updateCredentials(
-      user,
-      projectId,
-      dto,
-      request.auth?.token || '',
-    );
   }
 
   @UseGuards(ResourceGuard)
