+++ src/connection-request/dto/connection-request.dto.ts
@@ -426,11 +426,3 @@ export class ConnectionDecisionDto {
   @Type(() => DatabaseInfoDto)
   dbDetails?: DatabaseInfoDto;
 }
-
-export class UpdateCredentialsDto {
-  @ApiProperty({ type: DatabaseInfoDto })
-  @IsDefined()
-  @ValidateNested()
-  @Type(() => DatabaseInfoDto)
-  dbDetails!: DatabaseInfoDto;
-}
