+++ src/project/project.controller.ts
@@ -15,6 +15,7 @@ import {
   UseGuards,
   HttpCode,
   HttpStatus,
+  Patch,
 } from '@nestjs/common';
 import { ProjectService } from './project.service';
 import {
@@ -25,6 +26,7 @@ import {
   SettingsDto,
   SettingsResponseDto,
   UpdateProjectDto,
+  UpdateCredentialsDto,
 } from './dto';
 import {
   ApiBadRequestResponse,
@@ -448,6 +450,69 @@ export class ProjectController {
     @Param('projectId', ParseUUIDPipe) projectId: string,
   ) {
     return await this.projectService.getProjectSettings(projectId);
+  }
+
+  @UseGuards(ResourceGuard)
+  @Scopes('view', 'edit', 'connect')
+  @Patch(':projectId/credentials')
+  @ApiOperation({
+    summary: 'Update connection credentials and retry connection',
+  })
+  @HttpCode(HttpStatus.NO_CONTENT)
+  @ApiNoContentResponse({
+    description: 'Credentials updated and connection retried',
+  })
+  @ApiBadRequestResponse({
+    description: 'Invalid project data for database operation',
+    content: {
+      'application/json': {
+        schema: { $ref: getSchemaPath(GenericErrorResponseDto) },
+        example: {
+          statusCode: 400,
+          message: 'Invalid project data for database operation',
+          error: 'DatabaseError',
+        },
+      },
+    },
+  })
+  @ApiNotFoundResponse({
+    description: 'Project not found',
+    content: {
+      'application/json': {
+        schema: { $ref: getSchemaPath(GenericErrorResponseDto) },
+        example: {
+          statusCode: 404,
+          message: 'Requested resource could not be found',
+          error: 'DatabaseError',
+        },
+      },
+    },
+  })
+  @ApiInternalServerErrorResponse({
+    description: 'Unexpected database error',
+    content: {
+      'application/json': {
+        schema: { $ref: getSchemaPath(GenericErrorResponseDto) },
+        example: {
+          statusCode: 500,
+          message: 'Database is temporarily unavailable',
+          error: 'DatabaseError',
+        },
+      },
+    },
+  })
+  async updateCredentials(
+    @Req() request: Request,
+    @CurrentUser() user: CurrentUserInfo,
+    @Param('projectId', ParseUUIDPipe) projectId: string,
+    @Body() dto: UpdateCredentialsDto,
+  ) {
+    return await this.projectService.updateCredentials(
+      user,
+      projectId,
+      dto,
+      request.auth?.token || '',
+    );
   }
 
   // TODO: need to see if this is needed
